requirements.txt
requests
sentence-transformers
faiss-cpu
numpy
tqdm
python-dotenv
